scripts/flaky-agent/requirements.txt
requests>=2.31.0
urllib3>=2.0.0
azure-cosmos>=4.7.0
azure-identity>=1.17.0
